Redo tsconfig.json
Redo th tsconfig.json. This is now a copy from the psl-parser project.
The settings are stricter, any compile errors are fixed. Also this
project needs to go through a round of changing `return` to `return
null` in case the method returns and object but there are cases where
nothing is returned.