Add Publish Release GitHub Actions
Add the publish build github action to publish a new release. This is
also a copy from psl-parser, with some change on the variable that is
used for the authentication token.

Similar to the psl-parser project, temporarily set the organization to
@mischareitsma so that the flow can be tested. Before opening the PR to
the main repo, this will be changed to some final organization name.