Change lint and test frameworks
This commit introduces a massive change in some of the underlying
development related frameworks that are used. Following the guidelines
to rely on as less as possible external libraries the choice is made to
not use jest anymore, but the node test framework. Jest was not working
nice with ESModules either.

The change of test framework also comes with removal and change of
configuration. No jest config is require anymore, and the debug
configuration had to be adapted. Another change is the removal of the
two TypeScript config files. Now only one is required, for actually
building a package.

In addition the eslint package itself got deprecated, and now the
new eslint modules are used. The lint rules themselves also got changed,
and all the lint errors are resolved.